Доработка логов

+++ src/engine/gapi/directx12/Swapchain_DX.cpp
@@ -1,4 +1,5 @@
 #include "Swapchain_DX.h"
+#include "engine/gapi/utils/GAPILogHelpers.h"
 
 #if defined(Z_D3D12)
 namespace zzz::engine
@@ -184,7 +185,7 @@ namespace zzz::engine
 			device->CreateRenderTargetView(m_RenderTargets[i].Get(), nullptr, currentHandle);
 		}
 
-		DOut("[Swapchain_DX::OnResize] Successfully resized to {}x{} (BackBufferFormat: {}).", m_Size.GetWidth(), m_Size.GetHeight(), static_cast<uint32_t>(m_BackBufferFormat));
+		DOut("[Swapchain_DX::OnResize] Successfully resized to {}x{} (BackBufferFormat: {}).", m_Size.GetWidth(), m_Size.GetHeight(), m_BackBufferFormat);
 	}
 }
 #endif // Z_D3D12
